Compatibility with 7.1. Deleted BreedableSpec because that now exists in the mod as "!unbreedable". Repackaged Specs to specs. 1.4.2
