+++ gradle.properties
@@ -1,5 +1,5 @@
 modGroup=com.m1zark
-modVersion=1.4
+modVersion=1.4.2
 modBaseName=customspecs
-forgeVersion=1.12.2-14.23.4.2705
+forgeVersion=1.12.2-14.23.5.2809
 mcpVersion=snapshot_20171003
